Split deferred node traversal out from check.cpp
